+++ shake/__init__.py
@@ -53,5 +53,5 @@ from .wrappers import *
 NotAllowed = Forbidden
 redirect_to = redirect
 
-__version__ = '1.5.1'
+__version__ = '1.5.2'